backend/requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
sqlalchemy==2.0.44
alembic==1.14.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
PyJWT==2.10.1
bcrypt==5.0.0
openai==2.8.0
python-multipart==0.0.20
pillow==11.3.0
email-validator==2.3.0
requests==2.31.0
google-generativeai==0.7.1
